% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{imputation}
\alias{imputation}
\title{função de imputação
Recebe o método de imputação escolhido, o dado com lacunas e o tamanho da lacuna
Retorna os a série temporal imputada}
\usage{
imputation(method, dado, n)
}
\arguments{
\item{method}{metodo a ser usado na imputação}

\item{dado}{dado com lacunas a serem imputadas}

\item{n}{tamanho da lacuna}
}
\description{
função de imputação
Recebe o método de imputação escolhido, o dado com lacunas e o tamanho da lacuna
Retorna os a série temporal imputada
}
